requirements.txt
google-generativeai==0.4.1
tiktoken==0.7.0
streamlit==1.34.0